Made referenced.list tester-specific
