+++ rust/origen/src/core/tester.rs
@@ -4,15 +4,17 @@ use super::model::timesets::timeset::Timeset;
 use crate::core::dut::Dut;
 use crate::core::reference_files;
 use crate::generator::ast::{Attrs, Node};
-use crate::prog_gen::Model;
+use crate::prog_gen::{Model, PatternReferenceType};
 use crate::testers::{instantiate_tester, SupportedTester};
-use crate::utility::differ::Differ;
+use crate::utility::differ::{ASCIIDiffer, Differ};
 use crate::utility::file_utils::to_relative_path;
 use crate::{add_children, node, text, text_line, with_current_job};
 use crate::{Error, Result};
 use crate::{FLOW, TEST};
 use indexmap::IndexMap;
+use std::collections::HashSet;
 use std::env;
+use std::io::Write;
 use std::path::{Path, PathBuf};
 
 #[derive(Debug)]
@@ -518,6 +520,61 @@ impl Tester {
                             displayln!("");
                         }
                     }
+                    // Now write the pattern referenced list
+                    let mut referenced_patterns = HashSet::new();
+                    for pat in &model.patterns {
+                        if pat.reference_type == PatternReferenceType::All || pat.reference_type == PatternReferenceType::Origen {
+                            let _ = referenced_patterns.insert(pat.path.to_owned());
+                        }
+                    }
+                    let dir = gen.output_dir()?.join("test_program");
+                    if !dir.exists() {
+                        std::fs::create_dir_all(&dir)?;
+                    }
+                    let list = dir.join("referenced.list");
+                    let mut f = std::fs::File::create(&list)?;
+                    let mut pats: Vec<_> = referenced_patterns.into_iter().collect();
+                    pats.sort();
+                    for pat in pats {
+                        writeln!(&mut f, "{}", pat)?;
+                    }
+
+                    if let Ok(p) = to_relative_path(&list, None) {
+                        display!("Created: {}", p.display());
+                    } else {
+                        display!("Created: {}", list.display());
+                    }
+                    if let Some(ref_dir) = crate::STATUS.reference_dir() {
+                        let ref_list = ref_dir.join("referenced.list");
+                        display!(" - ");
+                        if ref_list.exists() {
+                            let mut differ = ASCIIDiffer::new(&ref_list, &list);
+                            differ.ignore_comments("#")?;
+                            if differ.has_diffs()? {
+                                if let Err(e) = reference_files::create_changed_ref(Path::new("referenced.list"), &list, &ref_list) {
+                                    log_error!("{}", e);
+                                }
+                                self.stats.changed_program_files += 1;
+                                display_redln!("Diffs found");
+                                let old = to_relative_path(&ref_list, None).unwrap_or(ref_list);
+                                let new = to_relative_path(&list, None).unwrap_or(list.to_owned());
+                                let diff_tool = std::env::var("ORIGEN_DIFF_TOOL").unwrap_or("tkdiff".to_string());
+                                displayln!("  {} {} {} &", &diff_tool, old.display(), new.display());
+                                display!("  origen save_ref referenced.list");
+                            } else {
+                                display_green!("No diffs");
+                            }
+                        } else {
+                            self.stats.new_program_files += 1;
+                            if let Err(e) = reference_files::create_new_ref(Path::new("referenced.list"), &list, &ref_list) {
+                                log_error!("{}", e);
+                            }
+                            display_cyanln!("New file");
+                            display!("  origen save_ref referenced.list");
+                        }
+                    }
+                    displayln!("");
+
                 } else {
                     log_debug!("No files generated by tester '{}", gen.name());
                 }
@@ -650,7 +707,7 @@ pub trait TesterAPI: std::fmt::Debug + Interceptor + TesterID + TesterAPIClone {
     /// and not patgen and vice versa, in that case they will return an empty vector.
     fn render_program(&mut self) -> crate::Result<(Vec<PathBuf>, Model)> {
         log_debug!("Tester '{}' does not implement render_program", &self.id());
-        Ok((vec![], Model::new()))
+        Ok((vec![], Model::new(self.id())))
     }
 
     /// The tester should implement this to return a differ instance which is configured
